requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
openpyxl==3.1.2
pydantic==2.5.0
requests==2.31.0